Test/1086 audit logger pagination boundary (#1146)
* test: verify execute_proposal timelock enforcement (#1078)

* test: verify audit_logger pagination boundaries (#1086)

---------

+++ onchain/contracts/audit_logger/tests/test_audit.rs
@@ -8,8 +8,6 @@ use soroban_sdk::{
     Address, Env, Symbol,
 };
 
-use audit_logger::{AuditLoggerContract, AuditLoggerContractClient, MAX_PAGE_SIZE};
-
 fn setup() -> (Env, Address, AuditLoggerContractClient<'static>) {
     let env = Env::default();
     env.mock_all_auths();
@@ -163,7 +161,7 @@ fn test_get_latest_logs_oversized_limit_clamped() {
     // Request far more than MAX_PAGE_SIZE.
     let latest = client.get_latest_logs(&u32::MAX);
     assert!(
-        latest.len() <= MAX_PAGE_SIZE,
+        latest.len() as u32 <= MAX_PAGE_SIZE,
         "get_latest_logs must clamp to MAX_PAGE_SIZE"
     );
 }
